erofs-utils: fix invalid argument type in erofs_err()
Coverity-id: 502378
Link: https://lore.kernel.org/r/20240805024408.2598464-1-hongzhen@linux.alibaba.com

+++ mkfs/main.c
@@ -397,7 +397,7 @@ static int parse_extended_opts(const char *opts)
 
 			if (!z_erofs_mkfs_features[i].feat) {
 				erofs_err("unknown extended option %.*s",
-					  p - token, token);
+					  (int)(p - token), token);
 				return -EINVAL;
 			}
 		}
